+++ src/Akeneo/Platform/Bundle/InstallerBundle/Command/DatabaseCommand.php
@@ -175,8 +175,8 @@ class DatabaseCommand extends Command
         $entityManager->clear();
 
         $this->eventDispatcher->dispatch(
-            InstallerEvents::POST_DB_CREATE,
-            new InstallerEvent($this->commandExecutor)
+            new InstallerEvent($this->commandExecutor),
+            InstallerEvents::POST_DB_CREATE
         );
 
         // TODO: Should be in an event subscriber
@@ -186,17 +186,17 @@ class DatabaseCommand extends Command
 
         if (false === $input->getOption('withoutFixtures')) {
             $this->eventDispatcher->dispatch(
-                InstallerEvents::PRE_LOAD_FIXTURES,
-                new InstallerEvent($this->commandExecutor)
+                new InstallerEvent($this->commandExecutor),
+                InstallerEvents::PRE_LOAD_FIXTURES
             );
 
             $this->loadFixturesStep($input, $output);
 
             $this->eventDispatcher->dispatch(
-                InstallerEvents::POST_LOAD_FIXTURES,
                 new InstallerEvent($this->commandExecutor, null, [
                     'catalog' => $input->getOption('catalog')
-                ])
+                ]),
+                InstallerEvents::POST_LOAD_FIXTURES
             );
         }
 
@@ -300,10 +300,10 @@ class DatabaseCommand extends Command
             ];
 
             $this->eventDispatcher->dispatch(
-                InstallerEvents::PRE_LOAD_FIXTURE,
                 new InstallerEvent($this->commandExecutor, $jobInstance->getCode(), [
                     'catalog' => $catalog
-                ])
+                ]),
+                InstallerEvents::PRE_LOAD_FIXTURE
             );
             if ($input->getOption('verbose')) {
                 $output->writeln(
@@ -315,8 +315,8 @@ class DatabaseCommand extends Command
             }
             $this->commandExecutor->runCommand('akeneo:batch:job', $params);
             $this->eventDispatcher->dispatch(
-                InstallerEvents::POST_LOAD_FIXTURE,
-                new InstallerEvent($this->commandExecutor, $jobInstance->getCode())
+                new InstallerEvent($this->commandExecutor, $jobInstance->getCode()),
+                InstallerEvents::POST_LOAD_FIXTURE
             );
         }
         $output->writeln('');