+++ src/Akeneo/Tool/Bundle/BatchBundle/spec/EventListener/LoggerSubscriberSpec.php
@@ -6,7 +6,7 @@ use Akeneo\Tool\Component\Batch\Event\JobExecutionEvent;
 use Akeneo\Tool\Component\Batch\Model\JobExecution;
 use PhpSpec\ObjectBehavior;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\Translation\TranslatorInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class LoggerSubscriberSpec extends ObjectBehavior
 {